P2/requirements.txt
django
Pillow
djangorestframework
djangorestframework-simplejwt